@@ -17,21 +17,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.0.0",
-    "@angular/common": "^12.0.0",
-    "@angular/compiler": "^12.0.0",
-    "@angular/core": "^12.0.0",
+    "@angular/animations": "^13.0.1",
+    "@angular/common": "^13.0.1",
+    "@angular/compiler": "^13.0.1",
+    "@angular/core": "^13.0.1",
     "@angular/fire": "../../dist/packages-dist",
-    "@angular/forms": "^12.0.0",
-    "@angular/platform-browser": "^12.0.0",
-    "@angular/platform-browser-dynamic": "^12.0.0",
-    "@angular/platform-server": "^12.0.0",
-    "@angular/router": "^12.0.0",
-    "@angular/service-worker": "^12.0.0",
-    "@nguniversal/express-engine": "^12.1.0",
+    "@angular/forms": "^13.0.1",
+    "@angular/platform-browser": "^13.0.1",
+    "@angular/platform-browser-dynamic": "^13.0.1",
+    "@angular/platform-server": "^13.0.1",
+    "@angular/router": "^13.0.1",
+    "@angular/service-worker": "^13.0.1",
+    "@nguniversal/express-engine": "^13.0.0",
     "cross-fetch": "^3.1.4",
     "express": "^4.15.2",
-    "firebase": "^9.0.0",
+    "firebase": "^9.4.0",
     "lodash.isequal": "^4.5.0",
     "rxfire": "^6.0.0",
     "rxjs": "~6.6.0",
@@ -40,11 +40,11 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1200.0",
-    "@angular-devkit/build-angular": "^12.0.0",
-    "@angular/cli": "^12.0.0",
-    "@angular/compiler-cli": "^12.0.0",
-    "@nguniversal/builders": "^12.1.0",
+    "@angular-devkit/architect": "~0.1300.2",
+    "@angular-devkit/build-angular": "^13.0.2",
+    "@angular/cli": "^13.0.2",
+    "@angular/compiler-cli": "^13.0.1",
+    "@nguniversal/builders": "^13.0.0",
     "@types/express": "^4.17.0",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
@@ -64,14 +64,11 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^12.0.0",
+    "ng-packagr": "^13.0.3",
     "open": "^7.0.3",
     "ts-node": "~9.1.1",
     "tslint": "~6.1.3",
-    "typescript": "~4.2.3",
+    "typescript": "~4.4.4",
     "webpack-bundle-analyzer": "^4.4.1"
-  },
-  "resolutions": {
-    "webpack": "^5.35.0"
   }
 }
