@@ -1,11 +1,8 @@
-import { getDatabase, connectDatabaseEmulator, ref, objectVal } from '@angular/fire/database';
-
-import { environment } from '../../environments/environment';
+import { getDatabase, ref, objectVal } from '@angular/fire/database';
+import { connectDatabaseEmulatorInDevMode } from '../emulators';
 
 const database = getDatabase();
-if (environment.useEmulators) {
-    connectDatabaseEmulator(database, 'localhost', 9000);
-}
+connectDatabaseEmulatorInDevMode(database);
 
 const doc = ref(database, 'test');
 export const valueChanges = objectVal(doc);
