@@ -1,10 +1,8 @@
-import { getFirestore, connectFirestoreEmulator, enableMultiTabIndexedDbPersistence } from '@angular/fire/firestore';
-import { environment } from '../environments/environment';
+import { getFirestore, enableMultiTabIndexedDbPersistence } from '@angular/fire/firestore';
+import { connectFirestoreEmulatorInDevMode } from './emulators';
 
 export const firestore = getFirestore();
-if (environment.useEmulators) {
-    connectFirestoreEmulator(firestore, 'localhost', 8080);
-}
+connectFirestoreEmulatorInDevMode(firestore);
 
 const isNode = () => !!(typeof process !== 'undefined' && process.versions?.node);
 
