@@ -1,25 +1,24 @@
-import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders, APP_INITIALIZER } from '@angular/core';
-import { Analytics as FirebaseAnalytics, isSupported } from 'firebase/analytics';
-import { ɵgetDefaultInstanceOf, ɵmemoizeInstance, ɵAngularFireSchedulers, VERSION } from '@angular/fire';
+import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders, APP_INITIALIZER, Injector } from '@angular/core';
+import { Analytics as FirebaseAnalytics } from 'firebase/analytics';
+import { ɵgetDefaultInstanceOf, ɵAngularFireSchedulers, VERSION, ɵisAnalyticsSupportedFactory } from '@angular/fire';
 import { Analytics, ANALYTICS_PROVIDER_NAME, AnalyticsInstances } from './analytics';
 import { FirebaseApps, FirebaseApp } from '@angular/fire/app';
 import { registerVersion } from 'firebase/app';
+import { ScreenTrackingService } from './screen-tracking.service';
+import { UserTrackingService } from './user-tracking.service';
 
-const PROVIDED_ANALYTICS_INSTANCES = new InjectionToken<Analytics[]>('angularfire2.analytics-instances');
-const IS_SUPPORTED = new InjectionToken<boolean>('angularfire2.analytics.isSupported');
+export const PROVIDED_ANALYTICS_INSTANCES = new InjectionToken<Analytics[]>('angularfire2.analytics-instances');
 
-const isSupportedSymbol = Symbol('angularfire2.analytics.isSupported');
-
-export function defaultAnalyticsInstanceFactory(isSupported: boolean, provided: FirebaseAnalytics[]|undefined, defaultApp: FirebaseApp) {
-  const defaultAnalytics = isSupported ?
-    ɵgetDefaultInstanceOf<FirebaseAnalytics>(ANALYTICS_PROVIDER_NAME, provided, defaultApp) :
-    undefined;
-  return new Analytics(defaultAnalytics);
+export function defaultAnalyticsInstanceFactory(provided: FirebaseAnalytics[]|undefined, defaultApp: FirebaseApp) {
+  if (!ɵisAnalyticsSupportedFactory.sync()) { return null; }
+  const defaultAnalytics = ɵgetDefaultInstanceOf<FirebaseAnalytics>(ANALYTICS_PROVIDER_NAME, provided, defaultApp);
+  return defaultAnalytics && new Analytics(defaultAnalytics);
 }
 
-export function analyticsInstanceFactory(fn: () => FirebaseAnalytics) {
-  return (zone: NgZone, isSupported: boolean) => {
-    const analytics = isSupported ? ɵmemoizeInstance<FirebaseAnalytics>(fn, zone) : undefined;
+export function analyticsInstanceFactory(fn: (injector: Injector) => FirebaseAnalytics) {
+  return (zone: NgZone, injector: Injector) => {
+    if (!ɵisAnalyticsSupportedFactory.sync()) { return null; }
+    const analytics = zone.runOutsideAngular(() => fn(injector));
     return new Analytics(analytics);
   };
 }
@@ -35,7 +34,6 @@ const DEFAULT_ANALYTICS_INSTANCE_PROVIDER = {
   provide: Analytics,
   useFactory: defaultAnalyticsInstanceFactory,
   deps: [
-    IS_SUPPORTED,
     [new Optional(), PROVIDED_ANALYTICS_INSTANCES ],
     FirebaseApp,
   ]
@@ -47,32 +45,33 @@ const DEFAULT_ANALYTICS_INSTANCE_PROVIDER = {
     ANALYTICS_INSTANCES_PROVIDER,
     {
       provide: APP_INITIALIZER,
-      useValue: () => isSupported().then(it => globalThis[isSupportedSymbol] = it),
+      useValue: ɵisAnalyticsSupportedFactory.async,
       multi: true,
-    },
+    }
   ]
 })
 export class AnalyticsModule {
-  constructor() {
+  constructor(
+    @Optional() _screenTrackingService: ScreenTrackingService,
+    @Optional() _userTrackingService: UserTrackingService,
+  ) {
     registerVersion('angularfire', VERSION.full, 'analytics');
   }
 }
 
-export function provideAnalytics(fn: () => FirebaseAnalytics): ModuleWithProviders<AnalyticsModule> {
+export function provideAnalytics(fn: (injector: Injector) => FirebaseAnalytics, ...deps: any[]): ModuleWithProviders<AnalyticsModule> {
   return {
     ngModule: AnalyticsModule,
     providers: [{
-      provide: IS_SUPPORTED,
-      useFactory: () => globalThis[isSupportedSymbol],
-    }, {
       provide: PROVIDED_ANALYTICS_INSTANCES,
       useFactory: analyticsInstanceFactory(fn),
       multi: true,
       deps: [
         NgZone,
-        IS_SUPPORTED,
+        Injector,
         ɵAngularFireSchedulers,
         FirebaseApps,
+        ...deps,
       ]
     }]
   };
