@@ -1,67 +1,66 @@
-import { isDevMode, NgZone, Version } from '@angular/core';
+import { Version } from '@angular/core';
 import { FirebaseApp, getApps } from 'firebase/app';
 import { ComponentContainer } from '@firebase/component';
+import { isSupported as isRemoteConfigSupported } from 'firebase/remote-config';
+import { isSupported as isMessagingSupported } from 'firebase/messaging';
+import { isSupported as isAnalyticsSupported } from 'firebase/analytics';
 
 export const VERSION = new Version('ANGULARFIRE2_VERSION');
+
+const isAnalyticsSupportedValueSymbol = '__angularfire_symbol__analyticsIsSupportedValue';
+const isAnalyticsSupportedPromiseSymbol = '__angularfire_symbol__analyticsIsSupported';
+const isRemoteConfigSupportedValueSymbol = '__angularfire_symbol__remoteConfigIsSupportedValue';
+const isRemoteConfigSupportedPromiseSymbol = '__angularfire_symbol__remoteConfigIsSupported';
+const isMessagingSupportedValueSymbol = '__angularfire_symbol__messagingIsSupportedValue';
+const isMessagingSupportedPromiseSymbol = '__angularfire_symbol__messagingIsSupported';
+
+globalThis[isAnalyticsSupportedPromiseSymbol] ||= isAnalyticsSupported().then(it =>
+  globalThis[isAnalyticsSupportedValueSymbol] = it
+);
+
+globalThis[isMessagingSupportedPromiseSymbol] ||= isMessagingSupported().then(it =>
+  globalThis[isMessagingSupportedValueSymbol] = it
+);
+
+globalThis[isRemoteConfigSupportedPromiseSymbol] ||= isRemoteConfigSupported().then(it =>
+  globalThis[isRemoteConfigSupportedValueSymbol] = it
+);
+
+const isSupportedError = (module: string) =>
+  `The APP_INITIALIZER that is "making" isSupported() sync for the sake of convenient DI has not resolved in this
+context. Rather than injecting ${module} in the constructor, first ensure that ${module} is supported by calling
+\`await isSupported()\`, then retrieve the instance from the injector manually \`injector.get(${module})\`.`;
+
+export const ɵisMessagingSupportedFactory = {
+  async: () => globalThis[isMessagingSupportedPromiseSymbol],
+  sync: () => {
+    const ret = globalThis[isMessagingSupportedValueSymbol];
+    if (ret === undefined) { throw new Error(isSupportedError('Messaging')); }
+    return ret;
+  }
+};
+
+export const ɵisRemoteConfigSupportedFactory = {
+  async: () => globalThis[isRemoteConfigSupportedPromiseSymbol],
+  sync: () => {
+    const ret = globalThis[isRemoteConfigSupportedValueSymbol];
+    if (ret === undefined) { throw new Error(isSupportedError('RemoteConfig')); }
+    return ret;
+  }
+};
+
+export const ɵisAnalyticsSupportedFactory = {
+  async: () => globalThis[isAnalyticsSupportedPromiseSymbol],
+  sync: () => {
+    const ret = globalThis[isAnalyticsSupportedValueSymbol];
+    if (ret === undefined) { throw new Error(isSupportedError('Analytics')); }
+    return ret;
+  }
+};
 
 // TODO is there a better way to get at the internal types?
 interface FirebaseAppWithContainer extends FirebaseApp {
   container: ComponentContainer;
-}
-
-const IS_HMR = !!(module as any).hot;
-
-const log = (level: 'log'|'error'|'info'|'warn', ...args: any) => {
-  if (isDevMode() && typeof console !== 'undefined') {
-    console[level](...args);
-  }
-};
-
-export function ɵcacheInstance<T>(cacheKey: any, moduleName: string, appName: string, fn: () => T, deps: any): T {
-  const [, instance, cachedDeps] = globalThis.ɵAngularfireInstanceCache.find((it: any) => it[0] === cacheKey) || [];
-  if (instance) {
-    if (!matchDep(deps, cachedDeps)) {
-      log('error', `${moduleName} was already initialized on the ${appName} Firebase App with different settings.${IS_HMR ? ' You may need to reload as Firebase is not HMR aware.' : ''}`);
-      log('warn', {is: deps, was: cachedDeps});
-    }
-    return instance;
-  } else {
-    const newInstance = fn();
-    globalThis.ɵAngularfireInstanceCache.push([cacheKey, newInstance, deps]);
-    return newInstance;
-  }
-}
-
-globalThis.ɵAngularfireInstanceCache ||= [];
-
-// HACK HACK HACK
-// AppCheck stuff, here so we can get a jump on it. It's too late in the evaluation
-// if we do this in the app-check module. globalThis.ngDevMode allows me to test if
-// Angular is in DevMode before Angular initializes.
-if (typeof process !== 'undefined' && process.env?.FIREBASE_APPCHECK_DEBUG_TOKEN) {
-  globalThis.FIREBASE_APPCHECK_DEBUG_TOKEN ??= process.env.FIREBASE_APPCHECK_DEBUG_TOKEN;
-} else if (globalThis.ngDevMode) {
-  globalThis.FIREBASE_APPCHECK_DEBUG_TOKEN ??= true;
-}
-
-export function ɵmemoizeInstance<T>(fn: () => T, zone: NgZone): T {
-  const [, instance] = globalThis.ɵAngularfireInstanceCache.find((it: any) => matchDep(it[0], fn)) || [];
-  if (instance) {
-    return instance as T;
-  } else {
-    // TODO catch and add HMR warning
-    const instance = zone.runOutsideAngular(() => fn());
-    globalThis.ɵAngularfireInstanceCache.push([fn, instance]);
-    return instance;
-  }
-}
-
-function matchDep(a: any, b: any) {
-  try {
-    return a.toString() === b.toString();
-  } catch (_) {
-    return a === b;
-  }
 }
 
 export function ɵgetDefaultInstanceOf<T= unknown>(identifier: string, provided: T[]|undefined, defaultApp: FirebaseApp): T|undefined  {
@@ -75,7 +74,7 @@ export function ɵgetDefaultInstanceOf<T= unknown>(identifier: string, provided:
   // Grab the default instance from the defaultApp
   const defaultAppWithContainer: FirebaseAppWithContainer = defaultApp as any;
   const provider = defaultAppWithContainer.container.getProvider(identifier as never);
-  return provider.getImmediate();
+  return provider.getImmediate({ optional: true });
 }
 
 export const ɵgetAllInstancesOf = <T= unknown>(identifier: string, app?: FirebaseApp): Array<T> => {
