@@ -1,7 +1,7 @@
-import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders } from '@angular/core';
+import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders, Injector } from '@angular/core';
 import { Firestore as FirebaseFirestore } from 'firebase/firestore';
 import { AuthInstances  } from '@angular/fire/auth';
-import { ɵmemoizeInstance, ɵgetDefaultInstanceOf, ɵAngularFireSchedulers, VERSION } from '@angular/fire';
+import { ɵgetDefaultInstanceOf, ɵAngularFireSchedulers, VERSION } from '@angular/fire';
 import { Firestore, FirestoreInstances, FIRESTORE_PROVIDER_NAME } from './firestore';
 import { FirebaseApps, FirebaseApp } from '@angular/fire/app';
 import { registerVersion } from 'firebase/app';
@@ -11,14 +11,12 @@ export const PROVIDED_FIRESTORE_INSTANCES = new InjectionToken<Firestore[]>('ang
 
 export function defaultFirestoreInstanceFactory(provided: FirebaseFirestore[]|undefined, defaultApp: FirebaseApp) {
   const defaultFirestore = ɵgetDefaultInstanceOf<FirebaseFirestore>(FIRESTORE_PROVIDER_NAME, provided, defaultApp);
-  // TODO how do I throw if it's undefined, unless @Optional(), is there an Angular NULL_INJECTOR token
-  // or something, can I use an @NgModule providers or something?
-  return new Firestore(defaultFirestore);
+  return defaultFirestore && new Firestore(defaultFirestore);
 }
 
-export function firestoreInstanceFactory(fn: () => FirebaseFirestore) {
-  return (zone: NgZone) => {
-    const firestore = ɵmemoizeInstance<FirebaseFirestore>(fn, zone);
+export function firestoreInstanceFactory(fn: (injector: Injector) => FirebaseFirestore) {
+  return (zone: NgZone, injector: Injector) => {
+    const firestore = zone.runOutsideAngular(() => fn(injector));
     return new Firestore(firestore);
   };
 }
@@ -51,7 +49,7 @@ export class FirestoreModule {
   }
 }
 
-export function provideFirestore(fn: () => FirebaseFirestore): ModuleWithProviders<FirestoreModule> {
+export function provideFirestore(fn: (injector: Injector) => FirebaseFirestore, ...deps: any[]): ModuleWithProviders<FirestoreModule> {
   return {
     ngModule: FirestoreModule,
     providers: [{
@@ -60,11 +58,13 @@ export function provideFirestore(fn: () => FirebaseFirestore): ModuleWithProvide
       multi: true,
       deps: [
         NgZone,
+        Injector,
         ɵAngularFireSchedulers,
         FirebaseApps,
         // Firestore+Auth work better if Auth is loaded first
         [new Optional(), AuthInstances ],
         [new Optional(), AppCheckInstances ],
+        ...deps,
       ]
     }]
   };
