@@ -1,25 +1,22 @@
-import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders, APP_INITIALIZER } from '@angular/core';
-import { isSupported, Messaging as FirebaseMessaging } from 'firebase/messaging';
-import { ɵgetDefaultInstanceOf, ɵmemoizeInstance, ɵAngularFireSchedulers, VERSION } from '@angular/fire';
+import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders, Injector, APP_INITIALIZER } from '@angular/core';
+import { Messaging as FirebaseMessaging } from 'firebase/messaging';
+import { ɵgetDefaultInstanceOf, ɵAngularFireSchedulers, VERSION, ɵisMessagingSupportedFactory } from '@angular/fire';
 import { Messaging, MessagingInstances, MESSAGING_PROVIDER_NAME } from './messaging';
 import { FirebaseApps, FirebaseApp } from '@angular/fire/app';
 import { registerVersion } from 'firebase/app';
 
 const PROVIDED_MESSAGING_INSTANCES = new InjectionToken<Messaging[]>('angularfire2.messaging-instances');
-const IS_SUPPORTED = new InjectionToken<boolean>('angularfire2.messaging.isSupported');
 
-const isSupportedSymbol = Symbol('angularfire2.messaging.isSupported');
-
-export function defaultMessagingInstanceFactory(isSupported: boolean, provided: FirebaseMessaging[]|undefined, defaultApp: FirebaseApp) {
-  const defaultAuth = isSupported ?
-    ɵgetDefaultInstanceOf<FirebaseMessaging>(MESSAGING_PROVIDER_NAME, provided, defaultApp) :
-    undefined;
-  return new Messaging(defaultAuth);
+export function defaultMessagingInstanceFactory(provided: FirebaseMessaging[]|undefined, defaultApp: FirebaseApp) {
+  if (!ɵisMessagingSupportedFactory.sync()) { return null; }
+  const defaultMessaging = ɵgetDefaultInstanceOf<FirebaseMessaging>(MESSAGING_PROVIDER_NAME, provided, defaultApp);
+  return defaultMessaging && new Messaging(defaultMessaging);
 }
 
-export function messagingInstanceFactory(fn: () => FirebaseMessaging) {
-  return (zone: NgZone, isSupported: boolean) => {
-    const messaging = isSupported ? ɵmemoizeInstance<FirebaseMessaging>(fn, zone) : undefined;
+export function messagingInstanceFactory(fn: (injector: Injector) => FirebaseMessaging) {
+  return (zone: NgZone, injector: Injector) => {
+    if (!ɵisMessagingSupportedFactory.sync()) { return null; }
+    const messaging = zone.runOutsideAngular(() => fn(injector));
     return new Messaging(messaging);
   };
 }
@@ -35,7 +32,6 @@ const DEFAULT_MESSAGING_INSTANCE_PROVIDER = {
   provide: Messaging,
   useFactory: defaultMessagingInstanceFactory,
   deps: [
-    IS_SUPPORTED,
     [new Optional(), PROVIDED_MESSAGING_INSTANCES ],
     FirebaseApp,
   ]
@@ -47,7 +43,7 @@ const DEFAULT_MESSAGING_INSTANCE_PROVIDER = {
     MESSAGING_INSTANCES_PROVIDER,
     {
       provide: APP_INITIALIZER,
-      useValue: () => isSupported().then(it => globalThis[isSupportedSymbol] = it),
+      useValue: ɵisMessagingSupportedFactory.async,
       multi: true,
     },
   ]
@@ -58,21 +54,19 @@ export class MessagingModule {
   }
 }
 
-export function provideMessaging(fn: () => FirebaseMessaging): ModuleWithProviders<MessagingModule> {
+export function provideMessaging(fn: () => FirebaseMessaging, ...deps: any[]): ModuleWithProviders<MessagingModule> {
   return {
     ngModule: MessagingModule,
     providers: [{
-      provide: IS_SUPPORTED,
-      useFactory: () => globalThis[isSupportedSymbol],
-    }, {
       provide: PROVIDED_MESSAGING_INSTANCES,
       useFactory: messagingInstanceFactory(fn),
       multi: true,
       deps: [
         NgZone,
-        IS_SUPPORTED,
+        Injector,
         ɵAngularFireSchedulers,
         FirebaseApps,
+        ...deps,
       ],
     }]
   };
