@@ -1,20 +1,29 @@
-import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders } from '@angular/core';
+import { NgModule, Optional, NgZone, InjectionToken, ModuleWithProviders, PLATFORM_ID, Injector } from '@angular/core';
 import { FirebasePerformance } from 'firebase/performance';
-import { ɵgetDefaultInstanceOf, ɵmemoizeInstance, ɵAngularFireSchedulers, VERSION } from '@angular/fire';
+import { ɵgetDefaultInstanceOf, ɵAngularFireSchedulers, VERSION } from '@angular/fire';
 import { Performance, PerformanceInstances, PERFORMANCE_PROVIDER_NAME } from './performance';
 import { FirebaseApps, FirebaseApp } from '@angular/fire/app';
 import { registerVersion } from 'firebase/app';
+import { isPlatformBrowser } from '@angular/common';
 
 export const PROVIDED_PERFORMANCE_INSTANCES = new InjectionToken<Performance[]>('angularfire2.performance-instances');
 
-export function defaultPerformanceInstanceFactory(provided: FirebasePerformance[]|undefined, defaultApp: FirebaseApp) {
+export function defaultPerformanceInstanceFactory(
+  provided: FirebasePerformance[]|undefined,
+  defaultApp: FirebaseApp,
+  // tslint:disable-next-line:ban-types
+  platform: Object
+) {
+  if (!isPlatformBrowser(platform)) { return null; }
   const defaultPerformance = ɵgetDefaultInstanceOf<FirebasePerformance>(PERFORMANCE_PROVIDER_NAME, provided, defaultApp);
-  return new Performance(defaultPerformance);
+  return defaultPerformance && new Performance(defaultPerformance);
 }
 
-export function performanceInstanceFactory(fn: () => FirebasePerformance) {
-  return (zone: NgZone) => {
-    const performance = ɵmemoizeInstance<FirebasePerformance>(fn, zone);
+export function performanceInstanceFactory(fn: (injector: Injector) => FirebasePerformance) {
+  // tslint:disable-next-line:ban-types
+  return (zone: NgZone, platform: Object, injector: Injector) => {
+    if (!isPlatformBrowser(platform)) { return null; }
+    const performance = zone.runOutsideAngular(() => fn(injector));
     return new Performance(performance);
   };
 }
@@ -32,6 +41,7 @@ const DEFAULT_PERFORMANCE_INSTANCE_PROVIDER = {
   deps: [
     [new Optional(), PROVIDED_PERFORMANCE_INSTANCES ],
     FirebaseApp,
+    PLATFORM_ID,
   ]
 };
 
@@ -47,7 +57,7 @@ export class PerformanceModule {
   }
 }
 
-export function providePerformance(fn: () => FirebasePerformance): ModuleWithProviders<PerformanceModule> {
+export function providePerformance(fn: () => FirebasePerformance, ...deps: any[]): ModuleWithProviders<PerformanceModule> {
   return {
     ngModule: PerformanceModule,
     providers: [{
@@ -56,8 +66,11 @@ export function providePerformance(fn: () => FirebasePerformance): ModuleWithPro
       multi: true,
       deps: [
         NgZone,
+        PLATFORM_ID,
+        Injector,
         ɵAngularFireSchedulers,
         FirebaseApps,
+        ...deps,
       ]
     }]
   };
