@@ -21,7 +21,13 @@ function generateHostingConfig(project: string, dist: string, functionName: stri
       source: '*.[0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f][0-9a-f].+(css|js)',
       headers: [{
         key: 'Cache-Control',
-        value: 'public,max-age=31536000,immutable'
+        value: 'public,max-age=31536000,immutable',
+      }]
+    }, {
+      source: '/@(ngsw-worker.js|ngsw.json)',
+      headers: [{
+        key: 'Cache-Control',
+        value: 'no-cache',
       }]
     }],
     rewrites: [
