@@ -9,7 +9,7 @@ import * as glob from 'glob';
 
 // TODO infer these from the package.json
 const MODULES = [
-  'core', 'app', 'app-check', 'compat', 'analytics', 'auth', 'database', 'firestore', 'functions',
+  'core', 'app', 'app-check', 'auth-guard', 'compat', 'analytics', 'auth', 'database', 'firestore', 'functions',
   'remote-config', 'storage', 'messaging', 'performance', 'compat/analytics',
   'compat/auth-guard', 'compat/auth', 'compat/database', 'compat/firestore',
   'compat/functions', 'compat/remote-config', 'compat/storage', 'compat/messaging',
@@ -23,6 +23,7 @@ interface OverrideOptions {
   exportName?: string;
   zoneWrap?: boolean;
   blockUntilFirst?: boolean;
+  override?: boolean;
 }
 
 function zoneWrapExports() {
@@ -35,7 +36,7 @@ function zoneWrapExports() {
   ) => {
     const imported = await import(path);
     const toBeExported: Array<[string, string, boolean]> = exports.
-      filter(it => !it.startsWith('_') && overrides[it] !== null).
+      filter(it => !it.startsWith('_') && overrides[it] !== null && overrides[it]?.override !== true).
       map(importName => {
         const zoneWrap = typeof imported[importName] === 'function' &&
           (overrides[importName]?.zoneWrap ?? importName[0] !== importName[0].toUpperCase());
@@ -44,6 +45,7 @@ function zoneWrapExports() {
       });
     const zoneWrapped = toBeExported.filter(([, , zoneWrap]) => zoneWrap);
     const rawExport = toBeExported.filter(([, , zoneWrap]) => !zoneWrap);
+    const overridden = Object.keys(overrides).filter(key => overrides[key]?.override);
     await writeFile(join(process.cwd(), 'src', `${module}/${name}.ts`), `// DO NOT MODIFY, this file is autogenerated by tools/build.ts
 ${path.startsWith('firebase/') ? `export * from '${path}';\n` : ''}${
 zoneWrapped.length > 0 ? `import { ɵzoneWrap } from '@angular/fire';
@@ -54,17 +56,26 @@ import {
 export {
   ${rawExport.map(([importName, exportName]) => `${importName}${exportName === importName ? '' : `as ${exportName}`}`).join(',\n  ')}
 } from '${path}';
+` : ''}${overridden.length > 0 ? `
+export {
+  ${overridden.join(',\n  ')}
+} from './overrides';
 ` : ''}
 ${zoneWrapped.map(([importName, exportName]) => `export const ${exportName} = ɵzoneWrap(_${importName}, ${overrides[importName]?.blockUntilFirst ?? true});`).join('\n')}
 `); };
   return Promise.all([
-    reexport('analytics', 'firebase', 'firebase/analytics', tsKeys<typeof import('firebase/analytics')>()),
+    reexport('analytics', 'firebase', 'firebase/analytics', tsKeys<typeof import('firebase/analytics')>(), {
+      isSupported: { override: true },
+    }),
     reexport('app', 'firebase', 'firebase/app', tsKeys<typeof import('firebase/app')>()),
     reexport('app-check', 'firebase', 'firebase/app-check', tsKeys<typeof import('firebase/app-check')>()),
     reexport('auth', 'rxfire', 'rxfire/auth', tsKeys<typeof import('rxfire/auth')>()),
     reexport('auth', 'firebase', 'firebase/auth', tsKeys<typeof import('firebase/auth')>(), {
       debugErrorMap: null,
       inMemoryPersistence: null,
+      browserLocalPersistence: null,
+      browserSessionPersistence: null,
+      indexedDBLocalPersistence: null,
       prodErrorMap: null,
     }),
     reexport('database', 'rxfire', 'rxfire/database', tsKeys<typeof import('rxfire/database')>()),
@@ -80,6 +91,7 @@ ${zoneWrapped.map(([importName, exportName]) => `export const ${exportName} = ɵ
     reexport('functions', 'firebase', 'firebase/functions', tsKeys<typeof import('firebase/functions')>()),
     reexport('messaging', 'firebase', 'firebase/messaging', tsKeys<typeof import('firebase/messaging')>(), {
       onMessage: { blockUntilFirst: false },
+      isSupported: { override: true },
     }),
     reexport('remote-config', 'rxfire', 'rxfire/remote-config', tsKeys<typeof import('rxfire/remote-config')>(), {
       getValue: { exportName: 'getValueChanges' },
@@ -88,7 +100,9 @@ ${zoneWrapped.map(([importName, exportName]) => `export const ${exportName} = ɵ
       getBoolean: { exportName: 'getBooleanChanges' },
       getAll: { exportName: 'getAllChanges' },
     }),
-    reexport('remote-config', 'firebase', 'firebase/remote-config', tsKeys<typeof import('firebase/remote-config')>()),
+    reexport('remote-config', 'firebase', 'firebase/remote-config', tsKeys<typeof import('firebase/remote-config')>(), {
+      isSupported: { override: true },
+    }),
     reexport('storage', 'rxfire', 'rxfire/storage', tsKeys<typeof import('rxfire/storage')>(), {
       getDownloadURL: null,
       getMetadata: null,
